Fixed some compiler warnings

+++ include/cantera/transport/MultiTransport.h
@@ -138,8 +138,8 @@ public:
      *                   length = ldx * ndim
      */
     virtual void getSpeciesFluxes(size_t ndim, const doublereal* const grad_T,
-                                  int ldx,  const doublereal* const grad_X,
-                                  int ldf, doublereal* const fluxes);
+                                  size_t ldx,  const doublereal* const grad_X,
+                                  size_t ldf, doublereal* const fluxes);
 
     //! Get the molar diffusional fluxes [kmol/m^2/s] of the species, given the thermodynamic
     //! state at two nearby points.
